+++ projects/component-library/stories/components/Conversation/Conversation.stories.tsx
@@ -1,7 +1,7 @@
 import { Story } from '@storybook/react';
 
-import { Conversation, ConversationProps, Icon, ConversationItem } from './Conversation';
-import { ColorPair } from '../../../../helpers/media';
+import { Conversation, ConversationProps, ConversationItem } from './Conversation';
+import { ColorPair, Icon } from '../../../../helpers/media';
 
 const goodMorning: ConversationItem[] = [
   {
