+++ projects/component-library/stories/components/Conversation/Conversation.tsx
@@ -1,13 +1,7 @@
-import { ReactNode, Ref } from 'react';
-import { MdAccountCircle } from 'react-icons/md';
+import { Ref } from 'react';
 
 import { ConversationStyles as s } from './Conversation.styles';
-import { ColorPair } from '../../../../helpers/media';
-
-export enum Icon {
-  None,
-  Account,
-}
+import { ColorPair, renderIcon, Icon } from '../../../../helpers/media';
 
 export interface ConversationProps {
   chatName: string;
@@ -25,15 +19,6 @@ export interface ConversationItem {
   text: string;
   sentByMe: boolean;
 }
-
-const renderIcon = (icon: Icon): ReactNode => {
-  switch (icon) {
-    case Icon.None:
-      return;
-    case Icon.Account:
-      return <MdAccountCircle />;
-  }
-};
 
 export const Conversation = ({
   chatName,
