+++ projects/component-library/stories/components/HoverColor/HoverColor.tsx
@@ -1,7 +1,5 @@
-import { MdMail, MdLogout, MdRefresh, MdSearch, MdCircle } from 'react-icons/md';
-
 import { HoverColorStyles as s } from './HoverColor.styles';
-import { ColorPair } from '../../../../helpers/media';
+import { ColorPair, renderIcon, Icon } from '../../../../helpers/media';
 
 export interface HoverColorProps {
   text?: string;
@@ -10,29 +8,6 @@ export interface HoverColorProps {
   date?: string;
   colorPair?: ColorPair;
 }
-
-export enum Icon {
-  None,
-  Mail,
-  LogOut,
-  Refresh,
-  Search,
-}
-
-const renderIcon = (icon: Icon): JSX.Element => {
-  switch (icon) {
-    case Icon.None:
-      return <MdCircle />;
-    case Icon.Mail:
-      return <MdMail />;
-    case Icon.LogOut:
-      return <MdLogout />;
-    case Icon.Refresh:
-      return <MdRefresh />;
-    case Icon.Search:
-      return <MdSearch />;
-  }
-};
 
 export const HoverColor = ({
   text,
